Move betty.plugins.js_resource to betty.js_resources
